Fixed some bugs and deprecated non-const operator[] of Prob,Factor
* Fixed some bugs in the MatLab interface build system
* Fixed a bug in utils/fginfo.cpp
* Added GraphEL::GraphEL( const GraphAL& G ) constructor
* Improved factor.h/cpp:
  - Added get(size_t) and set(size_t,T) operators; get() is
    equivalent to "operator[](size_t) const" and set() should
    be used instead of the non-const operator[], which has been deprecated
* Improved prob.h/cpp:
  - Added get(size_t) and set(size_t,T) operators; get() is
    equivalent to "operator[](size_t) const" and set() should
    be used instead of the non-const operator[], which has been deprecated

+++ utils/fginfo.cpp
@@ -97,7 +97,7 @@ int main( int argc, char *argv[] ) {
         cout << "Pairwise interactions? " << fg.isPairwise() << endl;
         // Calculate treewidth using various heuristics, if requested
         if( calc_tw ) {
-            std::pair<size_t,size_t> tw;
+            std::pair<size_t,double> tw;
             tw = boundTreewidth(fg, &eliminationCost_MinNeighbors);
             cout << "Treewidth (MinNeighbors):     " << tw.first << " (" << tw.second << " states)" << endl;
             tw = boundTreewidth(fg, &eliminationCost_MinWeight);
